Replace /about with /work and /life, add a shared SiteNav
- Split the old /about page into /work (professional) and /life (personal)
- Extract SiteNav so the global nav stays constant on every page; pages with
  in-page sections render a thin sticky sub-bar with scroll-spy beneath it
- Redirect /about → /life and /mentalhealth → /web-development
- Update blog post footer to point at /work + /life instead of /about
