+++ next.config.js
@@ -1,4 +1,18 @@
 /** @type {import('next').NextConfig} */
 module.exports = {
   reactStrictMode: true,
-}
+  async redirects() {
+    return [
+      {
+        source: "/about",
+        destination: "/life",
+        permanent: true,
+      },
+      {
+        source: "/mentalhealth",
+        destination: "/web-development",
+        permanent: true,
+      },
+    ];
+  },
+};
